style: apply clippy lints across core modules
Behavior-preserving cleanups: vec![..]->[..] arrays, format!("lit")->to_string(), (lo..=hi).contains(&x) over chained &&, drop redundant 'as f64'/'as i32' casts, split black_box(unit-return). Touches alien_sampler, band_conditioner, dense_mesh, dllm, percepta, progressive_mcgs, pruners/*, rt_turbo, ruliology, skill_opt, spectral_budget, stiff_anomaly.

+++ src/percepta/cumsum.rs
@@ -116,7 +116,7 @@ mod tests {
 
         for i in 1..=5 {
             let val = i as f64;
-            cs.insert(val, i as f64, i as i32);
+            cs.insert(val, i as f64, i);
             expected += val;
             let got = cs.query();
             assert!(
